Getting more features into pypackage
