+++ apps/pypackage/tasks.py
@@ -1,7 +1,8 @@
 from celery.task import task
-from pypi.slurper import Slurper
+from pypackage.slurper import Slurper
 
 import xmlrpclib
+
 base_url = "http://pypi.python.org/pypi/"
 PYPI = xmlrpclib.Server(base_url)
 
@@ -31,3 +32,18 @@ def queue_all_pypi_packages(package_limit=None):
             print "Queued %s to slurp from PyPi" % package_name
         except UnicodeDecodeError, UnicodeError:
             print "Couldn't queue %s due to Unicode Error" % package_name
+
+@task
+def fetch_package_metadata(pk):
+    package = Package.objects.get(pk=pk)
+    package.fetch_metadata()
+    print "Updated %s's metadata" % package.slug
+
+@task
+def queue_package_updates(package_limit=None):
+    all_packages = Package.objects.all().values_list('id', flat=True)[:package_limit]
+
+    # If this is actually a performance issue, make it a map
+    for package in all_packages:
+        print "Queued id: %s for updating" % package
+        fetch_package_metadata.delay(package)
